refactor: got rid of additional resolver class
