--- wandb/integration/diffusers/resolvers/__init__.py
@@ -1,9 +1,0 @@
-from .multimodal import (
-    SUPPORTED_MULTIMODAL_PIPELINES,
-    DiffusersMultiModalPipelineResolver,
-)
-
-__all__ = [
-    "SUPPORTED_MULTIMODAL_PIPELINES",
-    "DiffusersMultiModalPipelineResolver",
-]
